+++ frontend/apps/mobile/app/(app)/org/[id]/reviews.tsx
@@ -1,4 +1,3 @@
-import { Image } from "expo-image";
 import { useLocalSearchParams, useRouter } from "expo-router";
 import { useQueries } from "@tanstack/react-query";
 import {
@@ -12,7 +11,6 @@ import { useMemo, useState } from "react";
 import {
   ActivityIndicator,
   ScrollView,
-  Text,
   TouchableOpacity,
   View,
 } from "react-native";
@@ -23,85 +21,10 @@ import { ThemedText } from "@/components/themed-text";
 import { ThemedView } from "@/components/themed-view";
 import { RatingAggregateCard } from "@/components/ReviewAggregate";
 import { FilterTabs } from "@/components/SortingButtons";
-import { RATING_OPTIONS } from "@/constants/ratings";
-import { AppColors, Colors } from "@/constants/theme";
+import { EventRatingCard } from "@/components/EventRatingCard";
+import { Colors } from "@/constants/theme";
 
 type SortValue = "most_rated" | "highest" | "lowest";
-
-function EventRatingCard({
-  event,
-  aggregate,
-  onPress,
-}: {
-  event: Event;
-  aggregate: ReviewAggregate | null;
-  onPress: () => void;
-}) {
-  const { t: translate } = useTranslation();
-  const avg = aggregate?.average_rating ?? 0;
-  const total = aggregate?.total_reviews ?? 0;
-  const match = RATING_OPTIONS.find((r) => r.rating === Math.round(avg));
-
-  return (
-    <TouchableOpacity
-      activeOpacity={0.8}
-      onPress={onPress}
-      className="flex-row items-center rounded-2xl bg-white p-3 gap-3"
-      style={{
-        shadowColor: "#000",
-        shadowOpacity: 0.06,
-        shadowRadius: 8,
-        shadowOffset: { width: 0, height: 2 },
-        elevation: 2,
-      }}
-    >
-      <View
-        className="w-[72px] h-[72px] rounded-xl overflow-hidden"
-        style={{ backgroundColor: AppColors.imagePlaceholder }}
-      >
-        {event.presigned_url ? (
-          <Image
-            source={{ uri: event.presigned_url }}
-            style={{ width: "100%", height: "100%" }}
-            contentFit="cover"
-          />
-        ) : (
-          <View className="flex-1 items-center justify-center">
-            <IconSymbol name="photo" size={24} color={AppColors.mutedText} />
-          </View>
-        )}
-      </View>
-
-      <View className="flex-1">
-        <Text
-          className="text-[15px] font-nunito-bold mb-1"
-          style={{ color: AppColors.primaryText }}
-        >
-          {event.title}
-        </Text>
-        {total > 0 && match ? (
-          <View className="flex-row items-center gap-1.5">
-            <Image source={match.image} style={{ width: 18, height: 18 }} />
-            <Text
-              className="text-[13px] font-nunito"
-              style={{ color: AppColors.secondaryText }}
-            >
-              {translate(match.labelKey)}{" "}
-              <Text style={{ color: AppColors.subtleText }}>({total})</Text>
-            </Text>
-          </View>
-        ) : (
-          <Text
-            className="text-[13px] font-nunito"
-            style={{ color: AppColors.subtleText }}
-          >
-            {translate("review.noReviews")}
-          </Text>
-        )}
-      </View>
-    </TouchableOpacity>
-  );
-}
 
 export default function OrgReviewsPage() {
   const { id } = useLocalSearchParams<{ id: string }>();
